rename guide to project #2
