+++ src/index.js
@@ -2,7 +2,7 @@
 import './google-fonts-lato-400-700.css';
 
 import React from 'react';
-import selectPage from './select-guide.jsx';
+import selectPage from './select-page.jsx';
 
 /* eslint-disable global-require, import/no-unresolved */
 const pages = {
@@ -24,28 +24,34 @@ const pages = {
 export default class extends React.Component {
   constructor(props) {
     super(props);
-    const { sdk, guide } = props;
-    if (!sdk && guide) throw Error('[skygear-getstarted] Error: SDK not specified');
-    if (sdk && !pages[sdk]) throw Error(`[skygear-getstarted] No such SDK: ${sdk}`);
-    if (guide && !pages[sdk][guide]) throw Error(`[skygear-getstarted] No such guide: ${guide}`);
+    const { sdk, project } = props;
+    if (!sdk && project) {
+      throw Error('[skygear-getstarted] Error: SDK not specified');
+    }
+    if (sdk && !pages[sdk]) {
+      throw Error(`[skygear-getstarted] No such SDK: ${sdk}`);
+    }
+    if (project && !pages[sdk][project]) {
+      throw Error(`[skygear-getstarted] No such project: ${project}`);
+    }
     this.state = props;
   }
 
   get template() {
-    const { sdk, guide } = this.state;
-    if (pages[sdk] && pages[sdk][guide]) {
-      return pages[sdk][guide];
+    const { sdk, project } = this.state;
+    if (pages[sdk] && pages[sdk][project]) {
+      return pages[sdk][project];
     }
     return selectPage;
   }
 
   render() {
-    const { sdk, guide } = this.state;
+    const { sdk, project } = this.state;
     const self = this;
     return React.createElement(this.template, {
-      sdk, guide,
+      sdk, project,
       setSDK: ((targetSDK) => self.setState({ sdk: targetSDK })),
-      setGuide: ((targetGuide) => self.setState({ guide: targetGuide })),
+      setProject: ((targetProject) => self.setState({ project: targetProject })),
     });
   }
 }
