+++ src/pages/android-new.jade
@@ -1,1 +1,1 @@
-h2= 'Guide for '+ sdk +': '+ guide +' project.'
+h2= 'Guide for '+ sdk +': '+ project +' project.'
